% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{braintumors}
\alias{braintumors}
\title{Brain Tumor Data}
\format{A list of two items:
\describe{
  \item{\code{Y}}{Matrix of 356 continuous gene expression measurements for 43 brain tumors.}
  \item{\code{class}}{Vector giving tumor type for each tumor represented in \code{Y}}
  ...
}}
\source{
Data constructed as follows from the \code{denoiseR} package (\url{https://github.com/julierennes/denoiseR}): \cr
\code{library(devtools)} \cr
\code{install_github("julierennes/denoiseR")} \cr
\code{library(denoiseR)} \cr
\code{data(tumors)} \cr
\code{Y <- as.matrix(tumors[order(tumors[, ncol(tumors)]), -ncol(tumors)])} \cr
\code{class <- tumors[order(tumors[, ncol(tumors)]), ncol(tumors)]} \cr
\code{braintumors <- list("Y" = Y, "class" = class)}
}
\usage{
\code{data(braintumors)} \cr
\code{Y <- braintumors[["Y"]]} \cr
\code{class <- braintumors[["class"]]}
}
\description{
A dataset containing gene expression data for 43 tumors
}
\keyword{datasets}
