% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estM.R
\name{estM}
\alias{estM}
\title{Function for estimating unknown mean parameters}
\usage{
estM(Y, ett, tol = 1e-07)
}
\arguments{
\item{\code{Y}}{a matrix or three-way tensor of observed data}

\item{\code{ett}}{a list returned by the \code{estTune} function that contains the tuning parameter estimates to be used}
}
\value{
A list of unknown mean parameter estimates.
}
\description{
\code{estM} uses block coordinate descent to obtain a
LANOVA penalized estimate of a mean matrix or three-way tensor.
}
