% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estTuneTest.R
\name{estTuneTest}
\alias{estTuneTest}
\title{Function for estimating unknown tuning parameters and testing the appropriateness of
heavy-tailed C.}
\usage{
estTuneTest(Y, lowpen = FALSE)
}
\arguments{
\item{\code{Y}}{a matrix or tensor of observed data}

\item{\code{lowpen}}{a logical value indicating whether or not tuning parameters for
penalizing lower-order unknown mean parameters should be obtained. Tuning parameters for
penalizing lower-order unknown mean parameters, done when \code{lowpen=TRUE}, is only
supported for matrix and three-way tensor \code{Y}}
}
\value{
A list of tuning parameter estimates and a test statistic.
}
\description{
\code{estTuneTest} returns tuning parameter estimates and a test statistic for testing the
appropriateness of heavy tailed C given an observed matrix or tensor, Y.
}
