% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fMRI}
\alias{fMRI}
\title{fMRI Data}
\format{A list of two items:
\describe{
  \item{\code{Y}}{Three-way tensor.}
  \item{\code{map}}{Matrix of (x, y, z) locations for elements of \code{Y}}
  ...
}}
\usage{
fMRI
}
\description{
A dataset containing fMRI measurements for a single subject
}
\keyword{datasets}
