% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fusarium}
\alias{fusarium}
\title{Fusarium Data}
\format{A list of one item:
\describe{
  \item{\code{Y}}{Three-way tensor of disease ratings for 20 varieties of wheat infected with 7 strains of Fusarium head blight over 4 years, 1990-1993.}
  ...
}}
\source{
Fusarium data used and printed in \url{http://www.jstor.org/stable/2533660?seq=1#page_scan_tab_contents}, processed as follows: \cr
\cr
\code{fus <- read.table("fusarium.txt", header = FALSE) # Text file of data from paper} \cr
\code{Y <- array(NA, dim = c(4, 7, 20))} \cr
\code{for (i in 1:4) Y[i, , ] <- as.matrix(fus[(i - 1)*7 + 1:7, ])} \cr
\code{Y <- log((Y/100)/(1 - (Y/100)))} \cr
\code{fusarium <- list("Y" = Y)} \cr
}
\usage{
\code{data(fusarium)} \cr
\code{Y <- fusarium[["Y"]]}
}
\description{
A dataset containing severity of disease incidence ratings for 20 varieties of wheat infected with 7 strains of Fusarium head blight over 4 years, 1990-1993.
}
\keyword{datasets}
