% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tumors}
\alias{tumors}
\title{Tumor data}
\format{A list of two items:
\describe{
  \item{Y}{Matrix of 356 continuous gene expression measurements for 43 tumors.}
  \item{class}{Vector giving tumor type for each tumor represented in \code{Y}}
  ...
}}
\source{
\url{https://github.com/julierennes/denoiseR}
}
\usage{
tumors
}
\description{
A dataset containing gene expression data for 43 tumors.
}
\keyword{datasets}
